[REFAC] rename AND component file and update include paths; improve error handling in Input and Output components
