+++ include/Circuit.hpp
@@ -1,5 +1,6 @@
 #ifndef CIRCUIT_HPP
 #define CIRCUIT_HPP
+
 #include <string>
 #include <memory>
 #include <unordered_map>
@@ -13,6 +14,7 @@ namespace nts
     class Circuit {
         public:
             ~Circuit() = default;
+
             void addComponent(const std::string &type, const std::string &name);
             void linkComponents(const std::string &name1, const std::string &name2, std::size_t pin1, std::size_t pin2) const;
             void simulate() const;
