+++ src/specialComponents/Input.hpp
@@ -3,7 +3,7 @@
 
 #include <string>
 #include <memory>
-#include "../AComponent.hpp"
+#include "../../include/AComponent.hpp"
 
 namespace nts
 {
