+++ src/specialComponents/True.hpp
@@ -3,7 +3,8 @@
 
 #include <string>
 #include <memory>
-#include "../AComponent.hpp"
+#include "../../include/AComponent.hpp"
+#include <iostream>
 
 namespace nts
 {
